split out filesystem logic from state object, to allow core logic to be used client side
